backend/requirements.txt
# could be this or pyproject.toml
fastapi
uvicorn
qrcode
Pillow
eth-wake
requests